updated services for community
