+++ src/routes/AppRoutes.jsx
@@ -35,8 +35,13 @@ const ServicesForCommunityLayout = lazy(() =>
 const ServicesForCommunity = lazy(() =>
   import("../screens/services-for-community/page")
 );
+
+const Plantation = lazy(() =>
+  import("../screens/services-for-community/plantation/page")
+);
+
 const BloodDonation = lazy(() =>
-  import("../screens/services-for-community/blood-donation")
+  import("../screens/services-for-community/blood-donation/page")
 );
 
 const TeacherCourses = lazy(() =>
@@ -91,11 +96,6 @@ const ChatLayout = lazy(() => import("../screens/chat"));
 const ChatHome = lazy(() => import("../screens/chat/Home"));
 const ActiveChat = lazy(() => import("../screens/chat/ActiveChat"));
 const NotFound = lazy(() => import("../screens/NotFound"));
-
-// plantation routes
-
-const PlantationLayout = lazy(() => import("../screens/plantation/Layout"));
-const PlantationHome = lazy(() => import("../screens/plantation/Home"));
 
 // iery
 const IERYHome = lazy(() => import("../screens/main/iery"));
@@ -198,14 +198,12 @@ function AppRoutes() {
           <Route path="create-post" element={<CreatePost />} />
         </Route> */}
 
-        <Route path="plantation" element={<PlantationLayout />}>
-          <Route index element={<PlantationHome />} />
-        </Route>
         <Route
           path="services-for-community"
           element={<ServicesForCommunityLayout />}
         >
           <Route index element={<ServicesForCommunity />} />
+          <Route path="plantation" element={<Plantation />} />
           <Route path="blood-donation" element={<BloodDonation />} />
         </Route>
