--- src/screens/plantation/Layout.jsx
@@ -1,6 +1,0 @@
-import {  Outlet } from "react-router-dom";
-
-
-export default function PlantationLayout() {
-  return <Outlet />;
-}
